requirements.txt
dask==2024.8.2
ijson==3.4.0
litellm==1.75.9
numpy==2.3.2
pandas==2.3.2
protobuf==6.32.0
python_dateutil==2.9.0.post0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.5
sqlglot==26.9.0
sqlparse==0.5.3
streamlit==1.48.1
tiktoken==0.8.0
tqdm==4.66.5
google-auth==2.40.3
